Migrated DarkObservation class into its own file.
